% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBH.R
\name{plotBH}
\alias{plotBH}
\title{Benjamini - Hochberg plot}
\usage{
plotBH(pRFpar.list, FDRprob = 0.2)
}
\arguments{
\item{pRFpar.list}{output list of pRFpar}

\item{FDRprob}{FDR probability tolarance. Default 0.2}
}
\description{
Benjamini - Hochberg plot
}
