\name{pRF-package}
\alias{pRF-package}
\docType{package}
\title{
Permutation based approaches to Random Forest feature selection
}
\description{
Functions to estimate the statistical significance of the Increase in MSE and
Decrease in Gini Coefficient metrics of random forest feature importance.
The main functions are pRF and sigplot. See documentation on these functions
for information on how to use them.
}

\details{
\tabular{ll}{
Package: \tab pRF\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-23\cr
License: \tab GPL-3\cr
}

}
\author{

Ankur Chakravarthy

Maintainer: ankur.chakravarthy.10@ucl.ac.uk

}
\references{

The concept of permuting response variables is loosely based off 

Altmann A, Tolosi L, Sander O, Lengauer T. Permutation importance: a corrected
feature importance measure. Bioinformatics. 2010 May 15;26(10):1340-7. doi:
10.1093/bioinformatics/btq134. Epub 2010 Apr 12. PubMed PMID: 20385727.
}